requirements.txt
numpy
opencv-python
pillow
torch
torchvision
matplotlib
scikit-learn
seaborn
customtkinter
tcasl